A new popup to get connection parameters is added
Added ui for getting mysql connection parameters.
The UI will be launched the very first time the application is launched
or every time the application is launched depending on the remember
choice.

The password is saved as plain text for now. Will be replaced with AES
encryption at a later point.

The password will only be saved in case of users selecting the remember
option.